Extract Meilisearch search provider into spree_meilisearch gem
Meilisearch is an optional integration, so it no longer ships inside
spree_core: stores on the default Database provider stop carrying a
dependency they never use, and the gem can move at its own pace.

Core keeps the provider interface, the indexing jobs and the reindex
task. The document presenter moves with the provider — its shape is
Meilisearch's own — so the search_product_presenter injection point now
has no core default and the installed provider gem supplies it.

The old class names resolve for one release with a deprecation warning,
and raise a message naming the gem to install when it is absent.
