+++ docs/developer/deployment/environment_variables.mdx
@@ -111,7 +111,7 @@ After setting these, enable the provider and reindex:
 
 ```ruby
 # config/initializers/spree.rb
-Spree.search_provider = 'Spree::SearchProvider::Meilisearch'
+Spree.search_provider = 'SpreeMeilisearch::SearchProvider'
 ```
 
 <CodeGroup>
