+++ docs/integrations/search/meilisearch.mdx
@@ -5,7 +5,11 @@ description: "Set up Meilisearch in Spree for fast, typo-tolerant product search
 
 ## Overview
 
-[Meilisearch](https://www.meilisearch.com/) is a fast, open-source search engine that provides typo tolerance, relevance ranking, faceted filtering, and sub-50ms search responses. Spree includes a built-in Meilisearch search provider that replaces the default SQL-based search.
+[Meilisearch](https://www.meilisearch.com/) is a fast, open-source search engine that provides typo tolerance, relevance ranking, faceted filtering, and sub-50ms search responses. The `spree_meilisearch` gem provides a Meilisearch search provider that replaces the default SQL-based search.
+
+<Note>
+In Spree 5.x this provider shipped inside `spree_core` as `Spree::SearchProvider::Meilisearch`. From Spree 6.0 it lives in its own `spree_meilisearch` gem as `SpreeMeilisearch::SearchProvider`. The old class names keep working with a deprecation warning for one release. No reindex is needed when upgrading — the documents are unchanged.
+</Note>
 
 **When to use Meilisearch:**
 - Catalogs with 1,000+ products
@@ -52,7 +56,7 @@ For local development, Meilisearch runs without authentication by default — no
 ### 2. Add the gem to your Gemfile
 
 ```ruby
-gem 'meilisearch', '>= 0.28'
+gem 'spree_meilisearch'
 ```
 
 <CodeGroup>
@@ -83,7 +87,7 @@ MEILISEARCH_URL=http://localhost:7700
 
 ```ruby
 # config/initializers/spree.rb
-Spree.search_provider = 'Spree::SearchProvider::Meilisearch'
+Spree.search_provider = 'SpreeMeilisearch::SearchProvider'
 ```
 
 ### 5. Index your products
