+++ docs/plans/5.4-search-provider.md
@@ -1,6 +1,6 @@
 # Search Provider Interface
 
-**Status:** 5.4 portion shipped (`Spree::SearchProvider::Base/Database/Meilisearch`, PgSearch removed, `add_search_scope` removed); 6.0 metafield faceting pending. 2026-07-19: spree-starter / create-spree-app no longer provision Meilisearch by default — the Database provider is the out-of-the-box default everywhere; Meilisearch is a documented opt-in — installs: uncomment the compose blocks + set `MEILISEARCH_URL`; monorepo dev server: `SPREE_MEILISEARCH=1 pnpm server:dev` (chains `scripts/docker-compose.meilisearch.yml`); either way run `spree:search:reindex` once on first boot. See decisions.md.
+**Status:** 5.4 portion shipped (`Spree::SearchProvider::Base/Database/Meilisearch`, PgSearch removed, `add_search_scope` removed); 6.0 metafield faceting pending. 2026-07-19: spree-starter / create-spree-app no longer provision Meilisearch by default — the Database provider is the out-of-the-box default everywhere; Meilisearch is a documented opt-in — installs: uncomment the compose blocks + set `MEILISEARCH_URL`; monorepo dev server: `SPREE_MEILISEARCH=1 pnpm server:dev` (chains `scripts/docker-compose.meilisearch.yml`); either way run `spree:search:reindex` once on first boot. See decisions.md. **2026-08-12:** the Meilisearch provider left core for its own `spree_meilisearch` gem (`spree/providers/meilisearch`), as this plan always intended — `Spree::SearchProvider::Meilisearch` → `SpreeMeilisearch::SearchProvider` and the document presenter moved with it (`Spree::Dependencies.search_product_presenter` now has NO core default; the gem registers it, and `#search_presentation` raises a directive error without one). Core keeps the interface (`Base`/`Database`/`SearchResult`/`FiltersResult`/`CustomFieldSchema`), the indexing jobs and the reindex task; the `meilisearch` gem is no longer a core dependency. Old constants resolve with a deprecation warning through `Spree::SearchProvider.const_missing` until 6.1.
 **Target:** Spree 5.4 (provider + DB provider + PgSearch removal + add_search_scope removal + Meilisearch provider ✓), 6.0 (metafield faceting)
 **Depends on:** None for 5.4. ProductType (6.0-product-types.md) for metafield faceting in 6.0.
 **Author:** Damian + Claude
@@ -128,7 +128,7 @@ Consistent with existing `products_finder`, `products_sorter` pattern.
 - **Remove PgSearch scopes from Product** — `search_by_name` (trigram) removed. ILIKE on name is sufficient for the Database provider.
 - **Remove PgSearch scopes from Variant** — `search_by_sku`, `search_by_sku_or_options` replaced by ILIKE equivalents.
 - **Remove `pg_search` from gemspec.** Currently an optional dependency. No migration needed — it's a gem removal.
-- **Meilisearch is the recommended upgrade path.** For merchants who want typo tolerance, stemming, and relevance tuning, install `spree_search_meilisearch`. This gives them far more than PgSearch ever did.
+- **Meilisearch is the recommended upgrade path.** For merchants who want typo tolerance, stemming, and relevance tuning, install `spree_meilisearch`. This gives them far more than PgSearch ever did.
 
 ### Remove `add_search_scope` macro
 
@@ -139,7 +139,7 @@ Consistent with existing `products_finder`, `products_sorter` pattern.
 
 ### Meilisearch is the recommended external provider
 
-- **Ship `spree_search_meilisearch` as an official Spree gem.** Uses `meilisearch-ruby` (raw client), NOT `meilisearch-rails` (no model-level overrides).
+- **Ship `spree_meilisearch` as an official Spree gem.** Uses `meilisearch-ruby` (raw client), NOT `meilisearch-rails` (no model-level overrides).
 - **Index all products** (all statuses). Store API applies `.active` scope to the base AR scope before passing to the provider. Admin passes all products. The provider searches the full index, AR scope intersects for visibility.
 - **Meilisearch handles search + filtering + faceting in one call** — replaces both Ransack filtering and `FiltersAggregator` with Meilisearch's native capabilities. Adjusted facet counts come for free.
 - **Docs recommend Meilisearch** as the production search solution. The Database provider is for development and small catalogs.
@@ -293,7 +293,7 @@ module Spree
 end
 ```
 
-### Meilisearch Provider (`spree_search_meilisearch` gem)
+### Meilisearch Provider (`spree_meilisearch` gem)
 
 Uses `meilisearch-ruby` (raw HTTP client). No model-level overrides.
 
@@ -629,7 +629,7 @@ type ProductFilter = PriceRangeFilter | AvailabilityFilter | OptionFilter | Cate
 
 ### Phase 5: Meilisearch provider (official gem)
 
-- Create `spree_search_meilisearch` gem (official, maintained by Spree team)
+- Create `spree_meilisearch` gem (official, maintained by Spree team)
 - Uses `meilisearch-ruby` gem (raw client, no model-level overrides)
 - Indexes all products (all statuses) — storefront scopes handle visibility
 - Implements: search + filter + faceted aggregation with adjusted counts in one call
