+++ spree/common_spree_dependencies.rb
@@ -17,8 +17,6 @@ platforms :ruby do
   gem 'sqlite3', '>= 2.0'
 end
 
-gem 'meilisearch', '>= 0.28'
-
 group :test do
   gem 'parallel_tests'
   gem 'database_cleaner-active_record'
