--- spree/core/spec/presenters/spree/search_provider/product_presenter_spec.rb
@@ -1,312 +1,0 @@
-require 'spec_helper'
-
-module Spree
-  RSpec.describe SearchProvider::ProductPresenter do
-    let(:store) { @default_store }
-    let(:product) { create(:product, name: 'Test Shirt') }
-    let(:presenter) { described_class.new(product, store) }
-
-    describe '#call' do
-      subject(:documents) { presenter.call }
-
-      it 'returns an array of documents' do
-        expect(documents).to be_an(Array)
-        expect(documents).not_to be_empty
-      end
-
-      context 'document structure' do
-        subject(:doc) { documents.first }
-
-        it 'includes composite id and product_id' do
-          expect(doc[:id]).to start_with(product.prefixed_id)
-          expect(doc[:product_id]).to eq(product.prefixed_id)
-        end
-
-        it 'includes locale and currency' do
-          expect(doc[:locale]).to be_present
-          expect(doc[:currency]).to be_present
-        end
-
-        it 'includes flat name and description' do
-          expect(doc[:name]).to eq('Test Shirt')
-          expect(doc[:slug]).to eq(product.slug)
-        end
-
-        it 'includes flat price' do
-          expect(doc).to have_key(:price)
-        end
-
-        it 'includes stock status' do
-          expect(doc).to have_key(:in_stock)
-        end
-
-        it 'includes timestamps' do
-          expect(doc[:created_at]).to be_present
-          expect(doc[:updated_at]).to be_present
-        end
-
-        it 'includes category data as prefixed IDs' do
-          expect(doc[:category_ids]).to be_an(Array)
-        end
-
-        it 'includes option data as prefixed IDs' do
-          expect(doc[:option_type_ids]).to be_an(Array)
-          expect(doc[:option_value_ids]).to be_an(Array)
-        end
-      end
-
-      context 'with categories' do
-        let(:root_category) { create(:category, store: store, name: 'Apparel') }
-        let(:category) { create(:category, store: store, parent: root_category) }
-        let(:product) { create(:product, name: 'Test Shirt', categories: [category]) }
-
-        it 'indexes category_ids as prefixed IDs including ancestors' do
-          doc = documents.first
-          expect(doc[:category_ids]).to include(category.prefixed_id)
-          expect(doc[:category_ids]).to include(root_category.prefixed_id)
-          doc[:category_ids].each { |id| expect(id).to start_with('ctg_') }
-        end
-      end
-
-      context 'with nested categories' do
-        let(:root_category) { create(:category, store: store, name: 'Apparel') }
-        let(:parent) { create(:category, store: store, parent: root_category, name: 'Men') }
-        let(:child) { create(:category, store: store, parent: parent, name: 'Jackets') }
-        let(:product) { create(:product, name: 'Test Jacket', channels: [store.default_channel], categories: [child]) }
-
-        it 'indexes all ancestor category IDs so parent-level filtering works' do
-          doc = documents.first
-          expect(doc[:category_ids]).to include(child.prefixed_id)
-          expect(doc[:category_ids]).to include(parent.prefixed_id)
-          expect(doc[:category_ids]).to include(root_category.prefixed_id)
-        end
-      end
-
-      context 'with collections' do
-        let(:collection) { create(:collection, store: store) }
-
-        before { Spree::ProductCollection.create!(collection: collection, product: product) }
-
-        it 'indexes collection_ids as flat prefixed IDs (no ancestors)' do
-          doc = documents.first
-          expect(doc[:collection_ids]).to eq([collection.prefixed_id])
-          expect(doc[:collection_ids].first).to start_with('coll_')
-        end
-      end
-
-      context 'membership documents (manual sort)' do
-        let(:collection) { create(:collection, store: store) }
-
-        before { Spree::ProductCollection.create!(collection: collection, product: product, position: 4) }
-
-        it 'emits a collection membership doc carrying position + full base payload' do
-          member = documents.find { |d| d[:grouping_id] == collection.prefixed_id }
-
-          expect(member).to be_present
-          expect(member[:position]).to eq(4)
-          expect(member[:name]).to eq('Test Shirt')            # full base payload copied
-          expect(member[:product_id]).to eq(product.prefixed_id)
-          expect(member[:id]).to include(collection.prefixed_id)
-        end
-
-        it 'leaves base docs without a grouping_id/position' do
-          base = documents.reject { |d| d[:grouping_id] }
-
-          expect(base).not_to be_empty
-          base.each { |d| expect(d[:position]).to be_nil }
-        end
-      end
-
-      context 'category membership positions (subtree-MIN)' do
-        let(:parent) { create(:category, store: store, name: 'Men') }
-        let(:jackets) { create(:category, store: store, parent: parent, name: 'Jackets') }
-        let(:coats) { create(:category, store: store, parent: parent, name: 'Coats') }
-        let(:product) { create(:product, name: 'Test Jacket', channels: [store.default_channel], categories: [jackets, coats]) }
-
-        before do
-          Spree::ProductCategory.where(product: product, category: jackets).update_all(position: 5)
-          Spree::ProductCategory.where(product: product, category: coats).update_all(position: 2)
-        end
-
-        it 'emits a membership doc per ancestor-or-self, folding parent to the subtree minimum' do
-          jackets_doc = documents.find { |d| d[:grouping_id] == jackets.prefixed_id }
-          coats_doc = documents.find { |d| d[:grouping_id] == coats.prefixed_id }
-          parent_doc = documents.find { |d| d[:grouping_id] == parent.prefixed_id }
-
-          expect(jackets_doc[:position]).to eq(5)
-          expect(coats_doc[:position]).to eq(2)
-          expect(parent_doc[:position]).to eq(2) # MIN(5, 2) across the subtree
-        end
-      end
-
-      context 'with multiple markets' do
-        let!(:us_market) { store.default_market }
-        let!(:eu_market) { create(:market, store: store, name: 'EU', currency: 'EUR', default_locale: 'de', supported_locales: 'de,fr') }
-
-        # Product with prices in both currencies
-        let(:product) do
-          p = create(:product, name: 'Test Shirt')
-          create(:price, variant: p.default_variant, amount: 29.99, currency: 'EUR')
-          p
-        end
-
-        # Reload store to pick up new markets
-        let(:presenter) { described_class.new(product, store.reload) }
-
-        it 'creates one document per market × locale where product has a price' do
-          locales_currencies = documents.map { |d| [d[:locale], d[:currency]] }
-          expect(locales_currencies).to include(['en', 'USD'])
-          expect(locales_currencies).to include(['de', 'EUR'])
-          expect(locales_currencies).to include(['fr', 'EUR'])
-          expect(documents.size).to eq(3)
-        end
-
-        it 'each document has the correct currency and locale' do
-          usd_doc = documents.find { |d| d[:currency] == 'USD' }
-          eur_doc = documents.find { |d| d[:currency] == 'EUR' && d[:locale] == 'de' }
-
-          expect(usd_doc[:locale]).to eq('en')
-          expect(eur_doc[:locale]).to eq('de')
-        end
-
-        it 'each document has a unique composite id' do
-          ids = documents.map { |d| d[:id] }
-          expect(ids.uniq.size).to eq(ids.size)
-        end
-
-        it 'composite id includes product_id, locale, and currency' do
-          doc = documents.first
-          expect(doc[:id]).to match(/\A#{product.prefixed_id}_\w+_[A-Z]{3}\z/)
-        end
-
-        it 'includes the price for each currency' do
-          usd_doc = documents.find { |d| d[:currency] == 'USD' }
-          eur_doc = documents.find { |d| d[:currency] == 'EUR' }
-
-          expect(usd_doc[:price]).to be_present
-          expect(eur_doc[:price]).to eq(29.99)
-        end
-
-        it 'all documents share the same product_id' do
-          product_ids = documents.map { |d| d[:product_id] }.uniq
-          expect(product_ids).to eq([product.prefixed_id])
-        end
-      end
-
-      context 'with market but product has no price in that currency' do
-        let!(:us_market) { store.default_market }
-        let!(:eu_market) { create(:market, store: store, name: 'EU', currency: 'EUR', default_locale: 'de') }
-
-        # Product with USD price only — no EUR price
-        let(:product) { create(:product, name: 'USD Only Product') }
-        let(:presenter) { described_class.new(product, store.reload) }
-
-        it 'skips markets where product has no price' do
-          currencies = documents.map { |d| d[:currency] }
-          expect(currencies).to include('USD')
-          expect(currencies).not_to include('EUR')
-        end
-
-        it 'only creates documents for markets with matching prices' do
-          expect(documents.size).to eq(1)
-          expect(documents.first[:currency]).to eq('USD')
-        end
-      end
-
-      context 'translation fallback' do
-        let!(:us_market) { create(:market, store: store, name: 'US', currency: 'USD', default_locale: 'en') }
-        let!(:eu_market) { create(:market, store: store, name: 'EU', currency: 'EUR', default_locale: 'de') }
-
-        let(:product) do
-          p = create(:product, name: 'English Name')
-          create(:price, variant: p.default_variant, amount: 19.99, currency: 'EUR')
-          p
-        end
-
-        let(:presenter) { described_class.new(product, store.reload) }
-
-        it 'falls back to default locale name when translation is missing' do
-          de_doc = documents.find { |d| d[:locale] == 'de' }
-          expect(de_doc[:name]).to eq('English Name')
-        end
-      end
-
-      context 'with option values' do
-        let(:product) { create(:product_with_option_types) }
-
-        it 'includes option value prefixed IDs' do
-          doc = documents.first
-          doc[:option_value_ids].each do |ov_id|
-            expect(ov_id).to start_with('optval_')
-          end
-        end
-      end
-
-      context 'with searchable / sortable custom_fields' do
-        let!(:searchable_def) do
-          create(:custom_field_definition, :short_text_field, :searchable,
-                 namespace: 'custom', key: 'pinyin_name', label: 'Pinyin')
-        end
-        let!(:sortable_def) do
-          create(:custom_field_definition, :number_field, :sortable,
-                 namespace: 'custom', key: 'priority', label: 'Priority')
-        end
-        let!(:ignored_def) do
-          create(:custom_field_definition, :short_text_field,
-                 namespace: 'custom', key: 'internal_note', label: 'Note')
-        end
-
-        before do
-          product.set_custom_field(searchable_def, 'mao-tai')
-          product.set_custom_field(sortable_def, '10')
-          product.set_custom_field(ignored_def, 'secret')
-        end
-
-        it 'indexes searchable and sortable custom_field values as cf_* attributes' do
-          doc = documents.first
-          expect(doc['cf_custom_pinyin_name']).to eq('mao-tai')
-          expect(doc['cf_custom_priority']).to eq(10.0)
-          expect(doc).not_to have_key('cf_custom_internal_note')
-        end
-
-        it 'uses the same cf_* attributes on every market/locale document' do
-          create(:market, store: store, name: 'US', currency: 'USD', default_locale: 'en')
-          create(:market, store: store, name: 'EU', currency: 'EUR', default_locale: 'de')
-          create(:price, variant: product.master, amount: 9.99, currency: 'EUR')
-
-          docs = described_class.new(product.reload, store.reload).call
-          expect(docs.size).to be > 1
-
-          cf_keys = docs.first.keys.grep(/\Acf_/)
-          docs.each do |doc|
-            expect(doc.slice(*cf_keys)).to eq(docs.first.slice(*cf_keys))
-          end
-        end
-
-        it 'casts number custom_field values to Float' do
-          decimal_def = create(:custom_field_definition, :number_field, :sortable,
-                               namespace: 'custom', key: 'rating', label: 'Rating')
-          product.set_custom_field(decimal_def, '42.5')
-
-          doc = described_class.new(product.reload, store).call.first
-          expect(doc['cf_custom_rating']).to eq(42.5)
-        end
-
-        it 'omits registered keys when the product has no value' do
-          create(:custom_field_definition, :short_text_field, :searchable,
-                 namespace: 'custom', key: 'origin', label: 'Origin')
-
-          doc = described_class.new(product.reload, store).call.first
-          expect(doc).not_to have_key('cf_custom_origin')
-        end
-
-        it 'includes no cf_* attributes when no definitions are searchable or sortable' do
-          Spree::CustomFieldDefinition.destroy_all
-
-          doc = described_class.new(product.reload, store).call.first
-          expect(doc.keys.grep(/\Acf_/)).to be_empty
-        end
-      end
-    end
-  end
-end
