+++ spree/providers/meilisearch/Gemfile
@@ -1,0 +1,5 @@
+eval_gemfile('../../common_spree_dependencies.rb')
+
+gem 'spree', path: '../../'
+
+gemspec
