+++ spree/providers/meilisearch/spree_meilisearch.gemspec
@@ -1,0 +1,30 @@
+# encoding: UTF-8
+
+require_relative '../../core/lib/spree/core/version.rb'
+
+Gem::Specification.new do |s|
+  s.platform    = Gem::Platform::RUBY
+  s.name        = 'spree_meilisearch'
+  s.version     = Spree.version
+  s.authors     = ['Spark Solutions Sp. z o.o.', 'Vendo Connect Inc.']
+  s.email       = 'hello@spreecommerce.org'
+  s.summary     = 'Meilisearch product search for Spree eCommerce platform'
+  s.description = 'Instant product search, filtering and faceted navigation for Spree via Meilisearch'
+  s.homepage    = 'https://spreecommerce.org'
+  s.license     = 'BSD-3-Clause'
+
+  s.metadata = {
+    "bug_tracker_uri"   => "https://github.com/spree/spree/issues",
+    "changelog_uri"     => "https://github.com/spree/spree/releases/tag/v#{s.version}",
+    "documentation_uri" => "https://docs.spreecommerce.org/",
+    "source_code_uri"   => "https://github.com/spree/spree/tree/v#{s.version}",
+  }
+
+  s.required_ruby_version = '>= 3.2'
+
+  s.files        = Dir["{app,config,db,lib,vendor}/**/*", "Rakefile", "README.md"].reject { |f| f.match(/^spec/) && !f.match(/^spec\/fixtures/) }
+  s.require_path = 'lib'
+
+  s.add_dependency 'meilisearch', '>= 0.28'
+  s.add_dependency 'spree_core', ">= #{s.version}"
+end
